Fixed two warnings related to float->int conversion.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@1651 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ pxd/dataobjects/include/PXDDigit.h
@@ -91,12 +91,12 @@ namespace Belle2 {
     /** Get cell ID in u.
      * @return u ID of the cell.
      */
-    float getUCellID() const { return m_uCellID; }
+    short int getUCellID() const { return m_uCellID; }
 
     /** Get cell ID in v.
      * @return v ID of the cell.
      */
-    float getVCellID() const { return m_vCellID; }
+    short int getVCellID() const { return m_vCellID; }
 
     /** Get u coordinate of cell center.
      * @return u coordinate of cell center.
